pyproject.toml
[build-system]
requires = ["setuptools>=68,<77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rs3-plugin-altitude-agpl"
version = "0.1.0"
description = "RS3 Altitude enricher (AGPL) — ajoute altitude_m via service SRTM/IGN"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
license = { text = "AGPL-3.0-only" }
authors = [{ name = "Sebastien Edet" }]
dependencies = [
  "pandas>=2.1",
  "pyyaml>=6.0",
  "requests>=2.31",
  "scipy>=1.10",
]

[project.urls]
Homepage = "https://github.com/SebE585/rs3-plugin-altitude-agpl"
Issues = "https://github.com/SebE585/rs3-plugin-altitude-agpl/issues"

# 👉 Déclaration du plugin RS3
[project.entry-points."rs3.plugins"]
altitude = "rs3_plugin_altitude_agpl.plugin:AltitudeStage"


[tool.setuptools.packages.find]
where = ["src"]

# Inclure les YAML embarqués (config par défaut + fragments de schéma)
[tool.setuptools.package-data]
rs3_plugin_altitude_agpl = [
  "config/*.yaml",
  "schema_fragments/*.yaml",
]


[tool.setuptools]
license-files = []

[project.scripts]
rs3-altitude = "rs3_plugin_altitude_agpl.factory:main"

[project.optional-dependencies]
test = [
  "pytest>=7.0",
]